Specify FQDN of Cognito login to Route53
This changes the Route53 A Record for Cognito's login page to specify the FQDN of the login page, instead of a relative version.  That's important if you're in the situation where your Route53 base is something like example.com, but your analytics domain is analytics.example.com, because in that case your Cognito login needs to be login.analytics.example.com.

Prior to this patch the domain was going to Route53 A Record as just "login", in that example, and so the record that was built was for login.example.com (relative to R53 base).  Instead, now, it'll build login.analytics.example.com.

Note that Route53 FQDN spec needs a recordName with a trailing dot (an actual FQDN).  The Cognito customDomain piece actually gets screwed up if it has that trailing dot.  Surprising.