Many changes - now in line w/discussion on PR #86
Shrank container min-width to allow more flexibility with content shrinking.  Now set to 160px, but including padding the minimum page width before the browser just starts shrinking content is about 220px.

Allowed cards to grow to fit space, so now if one of the 4 little data points drops to the next line it'll show as centered and look nice.  Also increased card min width and matched units to the font unit - so, essentially, the data points can show up to 8 or 9 large characters across before content will overflow.

Imported the Menu icon from vue-material-design so it can be used for the page view/event selector.

Added names to page and event in the router so the names can be displayed.

Modified the demo banner to work better at small screen widths.  Components will not encroach on the close button.  The close button and right side of the box are scooted away from the right wall a little.  The github links can now flex to a top line and push the text down below when there's not enough room for both of them on one line.

The bar charts in view and events have overflow and width set on them as constraints.  I'm honestly not sure why that makes them perform more consistently at smaller sizes in Firefox.  The 300 px width is not a magic number - the important part is that it's smaller than expected chart width at screen widths larger than 700 px.  This random fix only makes a difference at screen widths larger than 700 px...  Tested in Firefox and Chromium.

Site selector and date selector are now in a flexbox just by themselves, so when the screen gets too small for both then they can split across two lines and take up the full line.  Looks nice I think.

There's a new flexbox for the settings buttons and refresh button.  Views/Event content setting is in there (when visible), refresh button, and theme settings (cog) button is in there.  It's separate from the site/date selector so the site/date selectors flow nicer, and to allow this one to flow up in narrow mode.

700 pixels wide is the point at which the display switches mode.

The normal left side views/events bar becomes invisible when the screen is narrower than 700 px now.

The views/event "Content" selector hamburger button becomes visible when the screen is narrower than 700 px.  This has an el-drawer implementing the popout window.  The worst part about this is that this code is similar across two files now, instead of just being implemented in one place: views/stats/index.vue and vies/app.vue

The settings button flexbox (content, refresh, cog button) moves up a line, above the site/date selector, and takes up the whole width, when screen is narrower than 700px.

The map now pushes down below the location stats listing so they can both take the full width.  I think the map is a really cool feature, and easy enough to scroll past if you don't want to see it.